Separated ResSkyService from ProductService
