+++ src/main/java/com/myretail/api/service/ProductService.java
@@ -24,12 +24,4 @@ public interface ProductService {
      */
     void updateById(ProductPriceDTO productPriceDTO) throws ProductNotFoundException;
 
-    /**
-     * Call the "internal" API to retrieve the name for a given product id
-     * @param id product id
-     * @return name of the product
-     * @throws JSONException if an object does not have the required attributes to retrieve the name
-     * @throws ProductNotFoundException if we cannot find any product information from the given id
-     */
-    String findProductNameById(int id) throws JSONException, ProductNotFoundException;
 }
